requirements.txt
appdirs==1.4.4
atomicwrites==1.4.0
attrs==20.2.0
black==20.8b1
click==7.1.2
colorama==0.4.4
iniconfig==1.1.1
Jinja2==2.11.2
marko==0.9.1
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.4
pathspec==0.8.0
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
python-frontmatter==0.5.0
PyYAML==5.3.1
regex==2020.10.28
six==1.15.0
toml==0.10.1
typed-ast==1.5.0
typing-extensions==3.7.4.3
